Refactor next.config.mjs to add image domains
