+++ next.config.mjs
@@ -1,4 +1,8 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {};
-
-export default nextConfig;
+const nextConfig = {
+    images: {
+      domains: ['images.unsplash.com', 'avatar.vercel.sh', "media.licdn.com"],
+    },
+  };
+  
+  export default nextConfig;
